Move tests into TestExtendedMetadataIntegration
